files/BabelLayer_project_files/requirements.txt
# Core
PyQt6==6.6.1
PyQt6-Qt6==6.6.1
PyQt6-sip==13.6.0

# Database
SQLAlchemy==2.0.23

# Auth
PyJWT==2.8.0
bcrypt==4.1.2

# Data
pandas==2.1.4
numpy==1.26.2
openpyxl==3.1.2
xmltodict==0.13.0

# Mapping / Analytics
scikit-learn==1.3.2
sentence-transformers==2.2.2

# Explanations / API
openai==1.6.1
requests==2.31.0

# Reporting
reportlab==4.0.7
matplotlib==3.8.2

# Config
python-dotenv==1.0.0

# Dev
pytest==7.4.3
